Time: 52 ms (5.09%), Space: 46.2 MB (8.34%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,18 +1,18 @@
 class Solution {
     public int mySqrt(int x) {
         if(x==0 || x==1) return x;
-        int left = 0;
-        int right = (x/2)+1;
-        if(right*right==x) return right;
+        long left = 0;
+        long right = (x/2)+1;
+        if(right*right==x) return (int)right;
         while(right-left>1){
             long mid = (right+left)/2;
             if(mid*mid == x) return (int)mid;
 
-            else if(mid*mid < x) left=(int)mid;
+            else if(mid*mid < x) left=mid;
             
-            else right = (int)mid;
+            else right = mid;
             System.out.println(mid);
         }
-        return left;
+        return (int)left;
     }
 }
